# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Virtual Care API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same host"
    },
    {
      "url": "https://agent.fly.dev",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
...
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/invoices/{id}": {
      "get": {
        "operationId": "getInvoice",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": ["invoices"],
        "description": "Get a specific invoice by ID.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalAmount": {
                      "type": "number"
                    },
                    "client": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "anyOf": [
                            {
                              "type": "null"
                            },
                            {
                              "type": "string"
                            }
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "userId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "email",
                        "createdAt",
                        "updatedAt",
                        "userId"
                      ],
                      "additionalProperties": false
                    },
                    "invoiceLines": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "type": "object",
                            "properties": {
                              "project": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "name": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string"
                                  },
                                  "updatedAt": {
                                    "type": "string"
                                  },
                                  "userId": {
                                    "type": "string"
                                  },
                                  "description": {
                                    "anyOf": [
                                      {
                                        "type": "null"
                                      },
                                      {
                                        "type": "string"
                                      }
                                    ]
                                  },
                                  "hourlyRate": {
                                    "anyOf": [
                                      {
                                        "type": "null"
                                      },
                                      {
                                        "type": "string"
                                      }
                                    ]
                                  },
                                  "clientId": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "name",
                                  "createdAt",
                                  "updatedAt",
                                  "userId",
                                  "description",
                                  "hourlyRate",
                                  "clientId"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": ["project"],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "projectId": {
                                "type": "string"
                              },
                              "hours": {
                                "type": "string"
                              },
                              "rate": {
                                "type": "string"
                              },
                              "amount": {
                                "type": "string"
                              },
                              "invoiceId": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "description",
                              "projectId",
                              "hours",
                              "rate",
                              "amount",
                              "invoiceId"
                            ],
                            "additionalProperties": false
                          }
                        ]
                      }
                    },
                    "id": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "enum": ["DRAFT"],
                          "type": "string"
                        },
                        {
                          "enum": ["SENT"],
                          "type": "string"
                        },
                        {
                          "enum": ["PAID"],
                          "type": "string"
                        }
                      ]
                    },
                    "clientId": {
                      "type": "string"
                    },
                    "dateFrom": {
                      "type": "string"
                    },
                    "dateTo": {
                      "type": "string"
                    },
                    "sentAt": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "paidAt": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "paidAmount": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "pdfUrl": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    }
                  },
                  "required": [
                    "totalAmount",
                    "client",
                    "invoiceLines",
                    "id",
                    "createdAt",
                    "updatedAt",
                    "userId",
                    "status",
                    "clientId",
                    "dateFrom",
                    "dateTo",
                    "sentAt",
                    "paidAt",
                    "paidAmount",
                    "pdfUrl"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          },
          "404": {
            "description": "Response for 404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "enum": ["Invoice not found"],
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteInvoice",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": ["invoices"],
        "description": "Delete a draft invoice and unlock associated time entries.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "enum": ["Invoice deleted successfully"],
                      "type": "string"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "enum": ["Only draft invoices can be deleted"],
                            "type": "string"
                          },
                          "cause": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "enum": ["api/invoice-not-deletable"],
                                "type": "string"
                              },
                              "detail": {
                                "enum": [
                                  "Invoice must be in DRAFT status to be deleted"
                                ],
                                "type": "string"
                              }
                            },
                            "required": ["code", "detail"],
                            "additionalProperties": false
                          }
                        },
                        "required": ["message", "cause"],
                        "additionalProperties": false
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          },
          "404": {
            "description": "Response for 404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "enum": ["Invoice not found"],
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invoices/{id}/pdf": {
      "post": {
        "operationId": "generateInvoicePdf",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": ["invoices"],
        "description": "Generate and get download link for invoice PDF.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pdfUrl": {
                      "type": "string"
                    },
                    "message": {
                      "enum": ["PDF generation initiated"],
                      "type": "string"
                    }
                  },
                  "required": ["pdfUrl", "message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          },
          "404": {
            "description": "Response for 404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "enum": ["Invoice not found"],
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UnauthorizedErr": {
        "type": "object",
        "required": ["type", "title"],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Unauthorized"]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}